atom module and requisites included, with improved Node.header development
The subsequences of the ordered header are incorporated, allowing extension at intermediate points in the ordered sequence.
The atom module from Deliberation RuleML replaces the explicit content model for the element Atom.
It is also necessary to include slots, repo, resl and polyadic terms as well as closure, slot cardinality, slot weights.
We bring in degree at the same time.
An extension is needed to allow the “effectful” per attribute to occur on the Rel element.
Documentation of included components is removed or minimized.

Signed-off-by: greenTara <taraathan@gmail.com>